+++ src/lib/tls/tls13/tls_client_impl_13.h
@@ -88,7 +88,7 @@ class Client_Impl_13 : public Channel_Impl_13 {
       bool prepend_ccs() override;
 
       using Channel_Impl_13::handle;
-      void handle(const Server_Hello_12& server_hello_msg);
+      void handle(const Server_Hello_12_Shim& server_hello_msg);
       void handle(const Server_Hello_13& server_hello_msg);
       void handle(const Hello_Retry_Request& hrr_msg);
       void handle(const Encrypted_Extensions& encrypted_extensions_msg);
