+++ src/lib/tls/tls13/tls_handshake_state_13.h
@@ -70,7 +70,7 @@ class BOTAN_TEST_API Handshake_State_13_Base {
       Client_Hello_13& store(Client_Hello_13 client_hello, bool from_peer);
       Client_Hello_12& store(Client_Hello_12 client_hello, bool from_peer);
       Server_Hello_13& store(Server_Hello_13 server_hello, bool from_peer);
-      Server_Hello_12& store(Server_Hello_12 server_hello, bool from_peer);
+      Server_Hello_12_Shim& store(Server_Hello_12_Shim server_hello, bool from_peer);
       Hello_Retry_Request& store(Hello_Retry_Request hello_retry_request, bool from_peer);
       Encrypted_Extensions& store(Encrypted_Extensions encrypted_extensions, bool from_peer);
       Certificate_Request_13& store(Certificate_Request_13 certificate_request, bool from_peer);
@@ -100,7 +100,7 @@ class BOTAN_TEST_API Handshake_State_13_Base {
       std::optional<Client_Hello_13> m_client_hello;
       std::optional<Client_Hello_12> m_client_hello_12;
       std::optional<Server_Hello_13> m_server_hello;
-      std::optional<Server_Hello_12> m_server_hello_12;
+      std::optional<Server_Hello_12_Shim> m_server_hello_12;
       std::optional<Hello_Retry_Request> m_hello_retry_request;
       std::optional<Encrypted_Extensions> m_encrypted_extensions;
       std::optional<Certificate_Request_13> m_certificate_request;
