+++ src/lib/tls/tls_magic.h
@@ -10,6 +10,7 @@
 
 #include <botan/strong_type.h>
 #include <botan/types.h>
+#include <array>
 #include <vector>
 
 //BOTAN_FUTURE_INTERNAL_HEADER(tls_magic.h)
@@ -93,6 +94,38 @@ using Transcript_Hash = std::vector<uint8_t>;
 
 /// @brief Used to derive the ticket's PSK from the resumption_master_secret
 using Ticket_Nonce = Strong<std::vector<uint8_t>, struct Ticket_Nonce_>;
+
+/**
+ * Magic values used to signal a downgrade request to TLS 1.1.
+ *
+ * RFC 8446 4.1.3:
+ *   TLS 1.3 has a downgrade protection mechanism embedded in the server's
+ *   random value. TLS 1.3 servers which negotiate TLS 1.2 or below in
+ *   response to a ClientHello MUST set the last 8 bytes of their Random
+ *   value specially in their ServerHello.
+ */
+constexpr uint64_t DOWNGRADE_TLS11 = 0x444F574E47524400;
+
+/**
+ * Magic values used to signal a downgrade request to TLS 1.2.
+ *
+ * RFC 8446 4.1.3:
+ *   TLS 1.3 has a downgrade protection mechanism embedded in the server's
+ *   random value. TLS 1.3 servers which negotiate TLS 1.2 or below in
+ *   response to a ClientHello MUST set the last 8 bytes of their Random
+ *   value specially in their ServerHello.
+ */
+constexpr uint64_t DOWNGRADE_TLS12 = 0x444F574E47524401;
+
+/** 
+ * RFC 8446 4.1.3:
+ *   For reasons of backward compatibility with middleboxes, the
+ *   HelloRetryRequest message uses the same structure as the ServerHello, but
+ *   with Random set to the special value of the SHA-256 of "HelloRetryRequest":
+ */
+constexpr std::array<uint8_t, 32> HELLO_RETRY_REQUEST_MARKER = {
+   0xCF, 0x21, 0xAD, 0x74, 0xE5, 0x9A, 0x61, 0x11, 0xBE, 0x1D, 0x8C, 0x02, 0x1E, 0x65, 0xB8, 0x91,
+   0xC2, 0xA2, 0x11, 0x16, 0x7A, 0xBB, 0x8C, 0x5E, 0x07, 0x9E, 0x09, 0xE2, 0xC8, 0xA8, 0x33, 0x9C};
 
 }  // namespace Botan::TLS
